Перенес логику login и logout в blueprint user, сломал тесты)
