+++ app/__init__.py
@@ -1,11 +1,11 @@
-from flask_login import LoginManager, login_user, logout_user
+from flask_login import LoginManager
 
 from app.db import Session
-from app.models import User
+from app.user.models import User
 from app.user.forms import LoginForm
 from app.user.views import blueprint as user_blueprint
 from app.edit.edit import edit
-from flask import Flask, abort, render_template, redirect, url_for, flash
+from flask import Flask, abort, render_template
 from flask import session
 import os
 
@@ -36,25 +36,5 @@ def create_app():
         if "userLogged" not in session or session["userLogged"] != username:
             abort(401)
         return f"Профиль {username}"
-
-    @app.route('/process-login', methods=['POST'])
-    def process_login():
-        form = LoginForm()
-        if form.validate_on_submit():
-            with Session() as session:
-                user = session.query(User).filter_by(login=form.username.data).first()
-            if user and user.check_password(form.password.data):
-                login_user(user, remember=form.remember_me.data)
-                flash('Вы успешно авторизовались')
-                return redirect(url_for('index'))
-
-        flash('Неверный логин или пароль')
-        return redirect(url_for('user.login'))
-
-    @app.route('/logout')
-    def logout():
-        logout_user()
-        flash('Вы успешно разлогинились')
-        return redirect(url_for('index'))
 
     return app
