+++ app/templates/user/login.html
@@ -18,7 +18,7 @@ Login page
 <div class="container">
     <div class="row">
         <div class="col-4">
-            <form action="{{ url_for('process_login') }}" method="post">
+            <form action="{{ url_for('user.process_login') }}" method="post">
                 {{ form.hidden_tag() }}
                 <div class="form-group">
                     {{ form.username.label }}
