+++ app/user/views.py
@@ -1,7 +1,9 @@
-from flask import Blueprint, render_template, redirect, url_for
-from flask_login import current_user
+from flask import Blueprint, render_template, redirect, url_for, flash
+from flask_login import current_user, login_user, logout_user
 
+from app import Session
 from app.user.forms import LoginForm, RegistrationForm
+from app.user.models import User
 
 
 blueprint = Blueprint('user', __name__, url_prefix='/users')
@@ -23,3 +25,24 @@ def registration():
     title = "Регистрация"
     reg_form = RegistrationForm()
     return render_template("user/registration.html", page_title=title, form=reg_form)
+
+
+@blueprint.route('/process-login', methods=['POST'])
+def process_login():
+    form = LoginForm()
+    if form.validate_on_submit():
+        with Session() as session:
+            user = session.query(User).filter_by(login=form.username.data).first()
+        if user and user.check_password(form.password.data):
+            login_user(user, remember=form.remember_me.data)
+            flash('Вы успешно авторизовались')
+            return redirect(url_for('index'))
+
+    flash('Неверный логин или пароль')
+    return redirect(url_for('user.login'))
+
+@blueprint.route('/logout')
+def logout():
+    logout_user()
+    flash('Вы успешно разлогинились')
+    return redirect(url_for('index'))
